+++ nxdrive/gui/application.py
@@ -209,6 +209,7 @@ class Application(QApplication):
         context.setContextProperty("updater", self.manager.updater)
         context.setContextProperty("ratio", self.ratio)
         context.setContextProperty("isFrozen", Options.is_frozen)
+        context.setContextProperty("WINDOWS", WINDOWS)
         context.setContextProperty("tl", Translator._singleton)
         context.setContextProperty(
             "nuxeoVersionText", f"{self.manager.app_name} {self.manager.version}"